+++ scripts/_hr_retrieval_script_carmenes.py
@@ -1316,8 +1316,8 @@ def main(planet_name, output_directory, additional_data_directory, mode, uncerta
             ],
             rayleigh_species=['H2', 'He'],
             continuum_opacities=['H2-H2', 'H2-He'],
-            opacity_mode='lbl',
-            lbl_opacity_sampling=4,
+            line_opacity_mode='lbl',
+            line_by_line_opacity_sampling=4,
             # Temperature profile parameters
             temperature_profile_mode='isothermal',
             temperature=planet.equilibrium_temperature,  # K